refactor(AutoML): refactored AutoML script structure
